Dry out duplicate code
